# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel CRUD API Documentation",
        "description": "API for managing users, posts, and comments",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарий по ID",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Комментарий успешно найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновить существующий комментарий",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удалить комментарий",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Комментарий успешно удален"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{post}/posts": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарии для поста",
                "operationId": "getCommentsByPost",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список комментариев успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Добавить комментарий к посту",
                "operationId": "storeCommentToPost",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий успешно добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{comment}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Ответить на комментарий",
                "operationId": "storeCommentToComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ответ успешно добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{user}/users/active": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарии к активным постам пользователя",
                "operationId": "getCommentsForActivePostsByUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список комментариев успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/my/all": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить все комментарии текущего аутентифицированного пользователя",
                "operationId": "getMyComments",
                "responses": {
                    "200": {
                        "description": "Список комментариев успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{comment}/replies": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить ответы на комментарий",
                "operationId": "getRepliesToComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ответов успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentResource": {
                "title": "Comment Resource",
                "description": "Comment resource model",
                "properties": {
                    "id": {
                        "description": "ID of the comment",
                        "type": "integer",
                        "format": "int64"
                    },
                    "body": {
                        "description": "The body of the comment",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreCommentRequest": {
                "title": "Store Comment Request",
                "description": "Request body for creating a comment",
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "description": "The body of the comment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateCommentRequest": {
                "title": "Update Comment Request",
                "description": "Request body for updating a comment",
                "properties": {
                    "body": {
                        "description": "The body of the comment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}